move constants module
